+++ source/sidero/base/containers/readonlyslice.d
@@ -657,13 +657,13 @@ nothrow @nogc:
 
 private:
     static struct LifeTime {
-        shared(int) refCount;
+        shared(ptrdiff_t) refCount;
         RCAllocator allocator;
         LiteralType original;
     }
 
     static struct Iterator {
-        shared(int) refCount;
+        shared(ptrdiff_t) refCount;
         RCAllocator allocator;
         LiteralType literal;
